Move TracingConfiguration into its own file.
